requirements.txt
# 數據處理
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0

# 音頻處理
librosa>=0.8.1
soundfile>=0.10.3

# 機器學習
scikit-learn>=1.0.0
tensorflow>=2.7.0  # 或 pytorch>=1.10.0

# 可視化
matplotlib>=3.5.0
seaborn>=0.11.2

# 實用工具
tqdm>=4.62.0
pyyaml>=6.0

# 開發工具
jupyter>=1.0.0
black>=21.12b0
pylint>=2.12.0
